Add handling of ACQ1102 to /usr/local/CARE/acq400t_test
The script was previously only aware of ACQ1001 and ACQ2106.
For the general case it needs to run the DULPC test on an ACQ1102 in normal
operation.
At present it defaults to the ULPC test resulting in insufficient
coverage.
In addition to this it only looks for an FMC in Site 1 of an ACQ1001.
This should be done for both sites on the ACQ1102.
The knob for the configuration bit /dev/gpio/BV_FMC does not exist for ACQ1102.
Values are mapped onto /dev/gpio/BV1/value and /dev/gpio/BV2/value.
BV1 for Site 1 and BV2 for Site 2. 1= FMC, 0 = ELF.

	modified:   CARE/acq400t_test