Bugfix block certain input arguments to getenv
Passing a char as second argument would implicitly cast to an int.
This was causing a bug in the body of the function.
Now it cannot be called with (const char*, char) arguments.
The compiled will disallow this.
If you want to pass 'X' (char) then pass "X" (string) instead.p